+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Upgrade to Edgegrid v13.2.0
 * Update project license file
 * Migrate to go 1.26.4
+* Standardize command status/failure messaging ([OK/FAIL]) and table rendering format.
 
 ## Version 0.6.0 (January 2026)